+++ src/main/main.c
@@ -2,7 +2,7 @@
  * Main program entry point - read the command line options, then perform
  * the appropriate actions.
  *
- * Copyright 2002-2008, 2010, 2012-2015, 2017, 2021, 2023 Andrew Wood
+ * Copyright 2002-2008, 2010, 2012-2015, 2017, 2021, 2023-2024 Andrew Wood
  *
  * License GPLv3+: GNU GPL version 3 or later; see `docs/COPYING'.
  */
@@ -278,7 +278,7 @@ int main(int argc, char **argv)
 	pv_state_delay_start_set(state, opts->delay_start);
 	pv_state_linemode_set(state, opts->linemode);
 	pv_state_bits_set(state, opts->bits);
-	pv_state_si_set(state, opts->si);
+	pv_state_decimal_units_set(state, opts->decimal_units);
 	pv_state_null_terminated_lines_set(state, opts->null_terminated_lines);
 	pv_state_skip_errors_set(state, opts->skip_errors);
 	pv_state_error_skip_block_set(state, opts->error_skip_block);